requirements.txt
ccxt
pandas
numpy
websockets
python-dotenv
flask
gunicorn